breaks store and router into different files
related #8
